+++ packages/bylaw_ecto_query/lib/bylaw/ecto/query/checks/date_datetime_mixed_comparisons.ex
@@ -97,9 +97,14 @@ defmodule Bylaw.Ecto.Query.Checks.DateDatetimeMixedComparisons do
   @spec validate(Bylaw.Ecto.Query.Check.operation(), Bylaw.Ecto.Query.Check.query(), opts()) ::
           Bylaw.Ecto.Query.Check.result()
   def validate(operation, query, opts) when is_list(opts) do
-    check_opts = CheckOptions.normalize!(opts, [:validate])
+    check_opts = CheckOptions.normalize!(opts, [:validate, :rules])
 
-    if CheckOptions.enabled?(check_opts) do
+    if CheckOptions.enabled_in_scope?(
+         check_opts,
+         :date_datetime_mixed_comparisons,
+         operation,
+         query
+       ) do
       operation
       |> issues(query)
       |> result()