+++ packages/bylaw_ecto_query/lib/bylaw/ecto/query/checks/deterministic_order.ex
@@ -88,9 +88,10 @@ defmodule Bylaw.Ecto.Query.Checks.DeterministicOrder do
   @spec validate(Bylaw.Ecto.Query.Check.operation(), Bylaw.Ecto.Query.Check.query(), opts()) ::
           Bylaw.Ecto.Query.Check.result()
   def validate(operation, query, opts) when is_list(opts) do
-    check_opts = CheckOptions.normalize!(opts, [:validate])
+    check_opts = CheckOptions.normalize!(opts, [:validate, :rules])
 
-    if CheckOptions.enabled?(check_opts) and ordered?(query) do
+    if CheckOptions.enabled_in_scope?(check_opts, :deterministic_order, operation, query) and
+         ordered?(query) do
       validate_ordered_query(operation, query)
     else
       :ok