+++ packages/bylaw_ecto_query/lib/bylaw/ecto/query/checks/half_open_temporal_intervals.ex
@@ -69,6 +69,7 @@ defmodule Bylaw.Ecto.Query.Checks.HalfOpenTemporalIntervals do
   alias Bylaw.Ecto.Query.CheckOptions
   alias Bylaw.Ecto.Query.Introspection
   alias Bylaw.Ecto.Query.Issue
+  alias Bylaw.Ecto.Query.RuleOptions
 
   @comparison_operators [:<, :<=, :>, :>=]
   @temporal_types [
@@ -107,7 +108,7 @@ defmodule Bylaw.Ecto.Query.Checks.HalfOpenTemporalIntervals do
   @spec validate(Bylaw.Ecto.Query.Check.operation(), Bylaw.Ecto.Query.Check.query(), opts()) ::
           Bylaw.Ecto.Query.Check.result()
   def validate(operation, query, opts) when is_list(opts) do
-    check_opts = CheckOptions.normalize!(opts, [:validate, :fields])
+    check_opts = CheckOptions.normalize!(opts, [:validate, :fields, :rules])
 
     if CheckOptions.enabled?(check_opts) do
       validate_enabled(operation, query, check_opts)
@@ -121,19 +122,51 @@ defmodule Bylaw.Ecto.Query.Checks.HalfOpenTemporalIntervals do
   end
 
   defp validate_enabled(operation, query, check_opts) do
-    case checked_fields(query, check_opts) do
-      [] ->
-        :ok
+    rules = rules!(check_opts)
 
-      fields ->
-        operation
-        |> issues(query, fields)
-        |> result()
+    operation
+    |> RuleOptions.matching_rules(query, rules)
+    |> Enum.flat_map(fn rule ->
+      case checked_fields(query, rule) do
+        [] -> []
+        fields -> issues(operation, query, fields)
+      end
+    end)
+    |> Enum.uniq_by(&{&1.check, &1.message, &1.meta})
+    |> result()
+  end
+
+  defp rules!(opts) do
+    if Keyword.has_key?(opts, :rules) do
+      reject_top_level_fields_with_rules!(opts)
+
+      RuleOptions.rules!(
+        Keyword.fetch!(opts, :rules),
+        :half_open_temporal_intervals,
+        [:fields],
+        &rule_payload!/1
+      )
+    else
+      [Map.merge(rule_payload!(opts), %{where: [], except: []})]
+    end
+  end
+
+  defp reject_top_level_fields_with_rules!(opts) do
+    if Keyword.has_key?(opts, :fields) do
+      raise ArgumentError,
+            "expected half_open_temporal_intervals to use rule-level :fields when :rules is provided"
+    end
+  end
+
+  defp rule_payload!(opts) do
+    case Keyword.fetch(opts, :fields) do
+      {:ok, fields} -> %{fields: {:ok, normalize_fields!(fields)}}
+      :error -> %{fields: :infer}
     end
   end
 
   defp checked_fields(query, opts) do
-    case {configured_fields(opts), Introspection.root_schema(query)} do
+    case {opts.fields, Introspection.root_schema(query)} do
       {{:ok, fields}, {:ok, schema}} ->
         schema_fields = Introspection.schema_fields(schema)
         Enum.filter(fields, &MapSet.member?(schema_fields, &1))
@@ -146,13 +179,6 @@ defmodule Bylaw.Ecto.Query.Checks.HalfOpenTemporalIntervals do
 
       {:infer, :unknown} ->
         []
-    end
-  end
-
-  defp configured_fields(opts) do
-    case Keyword.fetch(opts, :fields) do
-      {:ok, fields} -> {:ok, normalize_fields!(fields)}
-      :error -> :infer
     end
   end